initial updates to the synthesize page to use SdsManager worker
