+++ packages/webapp/src/pages/Synthesize/DataSynthesis/hooks/index.ts
@@ -2,14 +2,8 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export * from './commands'
 export * from './useContextKey'
 export * from './useGetSyntheticCsvContent'
-export * from './useNoisyCountThresholdChange'
-export * from './useNoisyCountThresholdTypeOptions'
 export * from './useOnRunGenerateAndEvaluate'
-export * from './useOversamplingTypeOptions'
-export * from './usePrivacyBudgetProfileOptions'
 export * from './useSelectedContextParametersOnChange'
-export * from './useSynthesisModeOptions'
-export * from './useUseSyntheticCountOptions'
+export * from './useSyntheticTableCommands'
