--- packages/webapp/src/pages/Synthesize/DataSynthesis/hooks/useNoisyCountThresholdTypeOptions.ts
@@ -1,19 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-import type { IDropdownOption } from '@fluentui/react'
-
-import { NoisyCountThresholdType } from '~models'
-
-const noisyCountThresholdTypeOptions = [
-	{ key: NoisyCountThresholdType.Fixed, text: NoisyCountThresholdType.Fixed },
-	{
-		key: NoisyCountThresholdType.Adaptive,
-		text: NoisyCountThresholdType.Adaptive,
-	},
-]
-
-export function useNoisyCountThresholdTypeOptions(): IDropdownOption[] {
-	return noisyCountThresholdTypeOptions
-}
