+++ packages/webapp/src/pages/Synthesize/DataSynthesis/hooks/useSyntheticTableCommands.ts
