+++ packages/webapp/src/pages/Synthesize/DataSynthesisParameters/DataSynthesisParameters.styles.ts
@@ -1,0 +1,14 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+import { Dropdown, SpinButton } from '@fluentui/react'
+import styled from 'styled-components'
+
+export const StyledDropdown = styled(Dropdown)`
+	width: 200px;
+`
+
+export const StyledSpinButton = styled(SpinButton)`
+	width: 200px;
+`
