+++ packages/webapp/src/pages/Synthesize/DataSynthesisResult/DataSynthesisResult.types.ts
@@ -1,0 +1,10 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+
+import type { ISynthesisInfo } from '~workers/types'
+
+export interface DataSynthesisResultProps {
+	selectedSynthesis: ISynthesisInfo
+}
